pyproject.toml
[project]
name = "payment-system-service"
version = "0.0.0"
description = "A simple payment system service"
authors = [{ name = "Dêvid Teófilo", email = "nnick4673@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.4.1",
    "psycopg>=3.2.6",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-django>=4.11.1",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = "."
DJANGO_SETTINGS_MODULE = "payment_system_service.settings"
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
run = 'uvicorn payment_system_service.asgi:application --reload'
test = 'pytest -x -vv'
